Move test files to src/test/java because of errors
